Add some validation for basepath results (but they still don't do anything)

+++ src/main/kotlin/net/alexanderkahn/longball/service/model/BasepathResult.kt
@@ -1,3 +1,3 @@
 package net.alexanderkahn.longball.service.model
 
-data class BasepathResult(val player: Int, val location: PlayLocation, val result: PlayResult)
+data class BasepathResult(val player: Long, val location: PlayLocation, val result: PlayResult)
